apps/pre-processing-service/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "skip"}
]
requires-python = ">=3.11,<3.14"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
fastapi = ">=0.116.1,<0.117.0"
uvicorn = ">=0.35.0,<0.36.0"
loguru = ">=0.7.3,<0.8.0"
pydantic-settings = ">=2.10.1,<3.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
asyncpg = ">=0.30.0,<0.31.0"
gunicorn = ">=23.0.0,<24.0.0"
requests = ">=2.32.5,<3.0.0"
bs4 = ">=0.0.2,<0.0.3"
selenium = ">=4.35.0,<5.0.0"
transformers = ">=4.56.0,<5.0.0"
numpy = ">=2.3.2,<3.0.0"
scikit-learn = ">=1.7.1,<2.0.0"
python-dotenv = ">=1.1.1,<2.0.0"
mecab-python3 = ">=1.0.10,<2.0.0"
httpx = ">=0.28.1,<0.29.0"
pyperclip = ">=1.9.0,<2.0.0"
pymysql = ">=1.1.2,<2.0.0"
sqlalchemy = ">=2.0.43,<3.0.0"
google = "^3.0.0"
google-auth-oauthlib = "^1.2.2"
google-api-python-client = "^2.181.0"
poetry-core=">=2.1.3,<3.0.0"
dbutils=">=3.1.2,<4.0.0"
onnxruntime = "^1.22.1"
openai = "^1.107.3"
aiohttp = "^3.12.15"
prometheus-client = "^0.23.1"
prometheus-fastapi-instrumentator = "^7.1.0"
boto3 = "^1.40.35"
pytesseract = "^0.3.13"
pillow = "^11.3.0"
deep-translator = "^1.11.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest = "^8.4"

